# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Q-Vuln Post-Quantum Security Scanner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  description?: string;
  language: string;
  riskScore?: number | null;
  lastScanAt?: string | null;
  createdAt: string;
  totalScans: number;
  criticalCount: number;
  alertCount: number;
  cleanCount: number;
}

export interface CreateProjectBody {
  name: string;
  description?: string;
  language: string;
  code: string;
}

export type ScanMode = (typeof ScanMode)[keyof typeof ScanMode];

export const ScanMode = {
  "scan-only": "scan-only",
  interactive: "interactive",
  proactive: "proactive",
  community: "community",
} as const;

export type ScanStatus = (typeof ScanStatus)[keyof typeof ScanStatus];

export const ScanStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export type FindingSeverity =
  (typeof FindingSeverity)[keyof typeof FindingSeverity];

export const FindingSeverity = {
  critical: "critical",
  alert: "alert",
  safe: "safe",
} as const;

export interface Finding {
  id: number;
  scanId: number;
  fileName: string;
  lineNumber: number;
  severity: FindingSeverity;
  algorithm: string;
  codeSnippet: string;
  nistReplacement?: string | null;
  nistStandard?: string | null;
  effortHours: number;
  explanation?: string | null;
}

export interface Scan {
  id: number;
  projectId: number;
  mode: ScanMode;
  status: ScanStatus;
  riskScore?: number | null;
  totalLines: number;
  criticalCount: number;
  alertCount: number;
  cleanCount: number;
  totalEffortHours: number;
  estimatedCost: number;
  executiveSummary?: string | null;
  findings?: Finding[];
  createdAt: string;
  completedAt?: string | null;
}

export type CreateScanBodyMode =
  (typeof CreateScanBodyMode)[keyof typeof CreateScanBodyMode];

export const CreateScanBodyMode = {
  "scan-only": "scan-only",
  interactive: "interactive",
  proactive: "proactive",
  community: "community",
} as const;

export interface CreateScanBody {
  projectId: number;
  mode: CreateScanBodyMode;
  code: string;
  language: string;
}

export type CommunityPostType =
  (typeof CommunityPostType)[keyof typeof CommunityPostType];

export const CommunityPostType = {
  article: "article",
  question: "question",
  "migration-story": "migration-story",
} as const;

export interface CommunityPost {
  id: number;
  type: CommunityPostType;
  title: string;
  content: string;
  authorName: string;
  language?: string | null;
  framework?: string | null;
  upvotes: number;
  downvotes: number;
  tags: string[];
  createdAt: string;
}

export type CreateCommunityPostBodyType =
  (typeof CreateCommunityPostBodyType)[keyof typeof CreateCommunityPostBodyType];

export const CreateCommunityPostBodyType = {
  article: "article",
  question: "question",
  "migration-story": "migration-story",
} as const;

export interface CreateCommunityPostBody {
  type: CreateCommunityPostBodyType;
  title: string;
  content: string;
  authorName: string;
  language?: string;
  framework?: string;
  tags?: string[];
}

export type VoteBodyDirection =
  (typeof VoteBodyDirection)[keyof typeof VoteBodyDirection];

export const VoteBodyDirection = {
  up: "up",
  down: "down",
} as const;

export interface VoteBody {
  direction: VoteBodyDirection;
}

export type LeaderboardEntryBadge =
  (typeof LeaderboardEntryBadge)[keyof typeof LeaderboardEntryBadge];

export const LeaderboardEntryBadge = {
  bronze: "bronze",
  silver: "silver",
  gold: "gold",
  "quantum-guardian": "quantum-guardian",
} as const;

export interface LeaderboardEntry {
  rank: number;
  authorName: string;
  totalScans: number;
  totalPosts: number;
  totalUpvotes: number;
  badge: LeaderboardEntryBadge;
  points: number;
}

export type ActivityItemSeverity =
  (typeof ActivityItemSeverity)[keyof typeof ActivityItemSeverity];

export const ActivityItemSeverity = {
  critical: "critical",
  alert: "alert",
  info: "info",
} as const;

export interface ActivityItem {
  id: number;
  description: string;
  timestamp: string;
  severity: ActivityItemSeverity;
}

export interface GlobalStats {
  totalReposScanned: number;
  totalVulnerabilitiesFound: number;
  totalLinesScanned: number;
  totalCommunityPosts: number;
  totalMigrationsAssisted: number;
  mostCommonAlgorithm: string;
  recentActivity: ActivityItem[];
}

export interface DemoRepo {
  slug: string;
  name: string;
  description: string;
  language: string;
  stars: number;
  riskScore: number;
  criticalCount: number;
  alertCount: number;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface GithubScanBody {
  /** Public GitHub repository URL */
  repoUrl: string;
}

export interface GithubFileResult {
  path: string;
  language: string;
  lines: number;
  criticalCount: number;
  alertCount: number;
}

export interface GithubScanResult {
  repoUrl: string;
  owner: string;
  repo: string;
  totalFiles: number;
  findings: Finding[];
  criticalCount: number;
  alertCount: number;
  cleanCount: number;
  riskScore: number;
  totalLines: number;
  totalEffortHours?: number | null;
  executiveSummary: string;
  fileResults: GithubFileResult[];
}

export type ListCommunityPostsParams = {
  type?: ListCommunityPostsType;
  language?: string;
  limit?: number;
};

export type ListCommunityPostsType =
  (typeof ListCommunityPostsType)[keyof typeof ListCommunityPostsType];

export const ListCommunityPostsType = {
  article: "article",
  question: "question",
  "migration-story": "migration-story",
  all: "all",
} as const;

export type GetLeaderboardParams = {
  limit?: number;
};