# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Q-Vuln Post-Quantum Security Scanner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /community/posts:
    get:
      operationId: listCommunityPosts
      tags: [community]
      summary: List community posts
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [article, question, migration-story, all]
        - name: language
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Community posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityPost"
    post:
      operationId: createCommunityPost
      tags: [community]
      summary: Create a community post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommunityPostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /community/posts/{id}/vote:
    post:
      operationId: voteCommunityPost
      tags: [community]
      summary: Vote on a community post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteBody"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /community/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [community]
      summary: Get top contributors leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
...
components:
  schemas:
...
    CommunityPost:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [article, question, migration-story]
        title:
          type: string
        content:
          type: string
        authorName:
          type: string
        language:
          type: string
          nullable: true
        framework:
          type: string
          nullable: true
        upvotes:
          type: integer
        downvotes:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - content
        - authorName
        - upvotes
        - downvotes
        - tags
        - createdAt
    CreateCommunityPostBody:
      type: object
      properties:
        type:
          type: string
          enum: [article, question, migration-story]
        title:
          type: string
        content:
          type: string
        authorName:
          type: string
        language:
          type: string
        framework:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - type
        - title
        - content
        - authorName
    VoteBody:
      type: object
      properties:
        direction:
          type: string
          enum: [up, down]
      required:
        - direction
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        authorName:
          type: string
        totalScans:
          type: integer
        totalPosts:
          type: integer
        totalUpvotes:
          type: integer
        badge:
          type: string
          enum: [bronze, silver, gold, quantum-guardian]
        points:
          type: integer
      required:
        - rank
        - authorName
        - totalScans
        - totalPosts
        - totalUpvotes
        - badge
        - points
...